Add files via upload
added objects to javascript and gave them the variable 'locations'. Then created a new function update(area) to use the locations. then used the update function to be able to call the objects. This was used to be able to declutter the other functions since they were all calling different objects.